+++ src/main.rs
@@ -14,7 +14,10 @@ use std::{
     fs::{read_to_string, OpenOptions},
     env,
     io::Write,
+    ffi::OsStr,
 };
+
+use walkdir::WalkDir;
 
 pub mod tokenizer;
 pub mod parser;
@@ -28,10 +31,19 @@ use emitter::Emitter;
 use error::{Error, throw};
 
 
+/// Enumerates the types of processes that the Boron compiler can execute.
+#[derive(Copy, Clone)]
+pub enum Process {
+    Lib,
+    Exe,
+    Build,
+}
+
+
 /// Provides an abstraction over CLI arguments.
 pub struct Args {
     filename: Option<String>,
-    lib: bool,
+    process: Process,
 }
 
 impl Args {
@@ -39,7 +51,7 @@ impl Args {
     pub fn new() -> Self {
         Self {
             filename: None,
-            lib: false,
+            process: Process::Exe,
         }
     }
 
@@ -48,19 +60,24 @@ impl Args {
         self.filename = Some(f);
     }
 
-    /// Marks this as a library (rather than an executable).
+    /// Marks this as a header file (rather than an executable).
     pub fn mark_lib(&mut self) {
-        self.lib = true;
+        self.process = Process::Lib;
     }
 
     /// Marks this as an executable.
     pub fn mark_exe(&mut self) {
-        self.lib = false;
+        self.process = Process::Exe;
+    }
+
+    /// Marks this as a directory build.
+    pub fn mark_build(&mut self) {
+        self.process = Process::Build;
     }
 
     /// Gets whether or not this is a library.
-    pub fn is_lib(&self) -> bool {
-        self.lib
+    pub fn get_process(&self) -> Process {
+        self.process
     }
 
     /// Gets the filename from the CLI args.
@@ -89,18 +106,28 @@ fn main() {
 
     for arg in env::args() {
         if arg.starts_with("--") {
-            if arg == "--lib" {
-                args.mark_lib();
-            } else if arg == "--exe" {
-                args.mark_exe();
-            } else {
-                throw(Error::UnexpectedCliFlag (arg));
+            match arg.as_str() {
+                "--lib" => args.mark_lib(),
+                "--exe" => args.mark_exe(), // NOTE: this is marked by default
+                "--build" => args.mark_build(),
+                _ => throw(Error::UnexpectedCliFlag (arg)),
             }
         } else {
             args.set_filename(arg);
         }
     }
     
+    
+
+    match args.get_process() {
+        Process::Lib => compile_lib(args),
+        Process::Exe => compile_exe(args),
+        Process::Build => build(args),
+    };
+}
+
+
+fn compile_lib(args: Args) {
     let code = match read_to_string(&args.get_filename()) {
         Ok(c) => c,
         Err(_) => throw(Error::CouldNotReadFile (args.get_filename())),
@@ -115,20 +142,12 @@ fn main() {
 
     let mut output_filename = args.get_filename();
     output_filename.truncate(output_filename.len() - 4);
+    output_filename.push_str(".h");
 
-    let code = match args.is_lib() {
-        true => {
-            output_filename.push_str(".h");
-            emitter.compile_lib(args.get_bare_filename(), expressions)
-        },
-        false => {
-            output_filename.push_str(".c");
-            emitter.compile_exe(expressions)
-        },
-    };
+    let output = emitter.compile_lib(args.get_bare_filename(), expressions);
 
     // Open a file for output
-    let mut output = match OpenOptions::new()
+    let mut output_file = match OpenOptions::new()
         .create(true)
         .write(true)
         .truncate(true)
@@ -138,8 +157,67 @@ fn main() {
         Err(_) => throw(Error::CouldNotCreate (output_filename.to_owned())),
     };
 
-    match output.write_all(code.as_bytes()) {
+    match output_file.write_all(output.as_bytes()) {
         Ok(_) => (),
         Err(_) => throw(Error::CouldNotWriteFile (output_filename.to_owned())),
+    }
+}
+
+
+fn compile_exe(args: Args) {
+    let code = match read_to_string(&args.get_filename()) {
+        Ok(c) => c,
+        Err(_) => throw(Error::CouldNotReadFile (args.get_filename())),
+    };
+
+    let mut tokenizer = Tokenizer::new(code);
+
+    let parser = Parser::new();
+    let expressions = parser.parse_all(&mut tokenizer);
+
+    let mut emitter = Emitter::new();
+
+    let mut output_filename = args.get_filename();
+    output_filename.truncate(output_filename.len() - 4);
+    output_filename.push_str(".c");
+
+    let output = emitter.compile_exe(expressions);
+
+    // Open a file for output
+    let mut output_file = match OpenOptions::new()
+        .create(true)
+        .write(true)
+        .truncate(true)
+        .open(output_filename.to_owned())
+    {
+        Ok(f) => f,
+        Err(_) => throw(Error::CouldNotCreate (output_filename.to_owned())),
+    };
+
+    match output_file.write_all(output.as_bytes()) {
+        Ok(_) => (),
+        Err(_) => throw(Error::CouldNotWriteFile (output_filename.to_owned())),
+    }
+}
+
+
+fn build(args: Args) {
+    // Walk the given directory
+    let mut filenames: Vec<String> = Vec::new();
+    for entry in WalkDir::new(&args.get_filename()) {
+        let entry = match entry {
+            Ok(e) => e,
+            Err(_) => throw(Error::CouldNotReadFile (args.get_filename())),
+        };
+        if entry.path().extension() == Some(OsStr::new("brn")) {
+            filenames.push(entry.path().display().to_string());
+        }
+    }
+
+    for filename in filenames {
+        let mut filename_args = Args::new();
+        filename_args.set_filename(filename);
+        filename_args.mark_lib();
+        compile_lib(filename_args);
     }
 }